homework

+++ src/student_stanislav_astafjev/lesson_5/Task_5/ArrayDefinitionFive.java
@@ -2,6 +2,6 @@ package student_stanislav_astafjev.lesson_5.Task_5;
 
 class ArrayDefinitionFive {
     public static void main(String[] args) {
-        double[] numbers = {1, 2, 3,4 ,5};
+        int[] numbers = {1, 2, 3,4 ,5};
     }
 }
